Add API endpoint for listing images in a repo
The path /v0/images/<path> returns a list of the images available for the
repository named by <path>. There are three forms of name:

 1. <registry>/<org>/<image>, e.g., quay.io/coreos/etcd
 2. <org>/<image>, e.g., weaveworks/scope (registry defaults to
    "index.docker.io")
 3. <image>, e.g., debian (org defaults to "library")

There are a few wrinkles:

* If you have private repos, you need to supply credentials to the
registry hosting those repos. You do so by supplying a credentials
files in the same format as `~/.docker/config.json` (bind-mount it in
Docker, or provide it as a secret in Kubernetes); and, give the
in-container path to the supplied file as the argument `--credentials`
to the binary.

* Since we want to be based on `scratch`, we have to supply root
certificates to the image. I've included those from Ubuntu
16.04. Another way to do this would be to compile them into the binary
(but, on balance, it'd be better to let the binary look for them in
the surrounding OS).

* In the docker registry API v1, looking up metadata involves a
request per image. It's possible that quay.io's own API improves on
that, but first things first.