Add history DB and log release events to it
This adds a release status to the output of the commands `fluxctl
service show` and `fluxctl service list`, and a new command `fluxctl
service history`.

Each service gets a state and an event log, kept in a `history.DB`. So
far the events are basically "release started" and "release finished".

A state change is an event, but not all events are state changes. So,
these are separate in the `history.DB` interface.