Correct several problems with fluxy-in-k8s and document
The image did not include ssh, which git wants to be present.

If you mount a secret to get the private key, it will have the wrong
permissions and ssh will ignore it. So, copy the key into the working
directory, and make sure it has the right permissions.

Elaborated instructions on making keys for trying it out with a
config.