Return blank Repository if tags lookup fails
Some image repositories don't seem to have an entry in the Docker
registry v1 API, and return a 404. For the minute, gloss over these by
just returning a "blank" image.

We could use the v2 API, but it does not record created times for
tags, so we'd lose the ability to order them. (There may be another
temporary solution available via backwards compatibility information
that is exposed in the v2 API).