Make automator and server depend on common Releaser
To vastly simplify the automator, it can just use the Release
procedure from flux.Service. Unfortunately, the server (implementation
of flux.Service) also refers to the automator, to provide the
Automate/Deautomate methods.

To break this circularity, I factored out a Releaser interface and
implementation. The automator and the server both rely on it, rather
than on each other.

Since there're a few methods (and fields) which are handy for both the
releaser and the server, these are put in a helper which can be used
by both.