Don't quote YAML strings
Apparently this is bad, because some YAML parsers choke on quoted
strings. Not quoting them does have the problem that things that look
like numbers (like certain hex-encoded hashes) will be interpreted as
numbers, upsetting the decoder or application. In those cases, the
value is supposed to be quo-- .. wait a second..