Work around quay.io auth
quay.io does not quote the `scope` value in its WWW-Authenticate
header, causing the library we use
(github.com/heroku/docker-registry-client) for the v2 registry API to
fail to parse the header.

The library authenticates every single request by default, meaning
that every manifest lookup (to get the image created time) uses three
roundtrips.

To solve both of these, this commit introduces a transport wrapper that

 1. rewrites the WWW-Authenticate header of responses if necessary;
 2. remembers the authentication bearer token (once established) for
 subsequent requests.

This commit fixes a silly bug, too: the auth entries in a docker
config are keyed by the host name, not "https://" + the host name.