Show release action results when executed
This adds a Result field to release actions, which is filled in as
each is executed. The main reason to support this is to show when an
action has no effect; e.g., when the config turns out to not be
present, and a service update is skipped.

It's also possible that there are no changes to config when it comes
to update it; for this scenario, the git CommitAndPush operation now
checks for differences first, and exits early if there are none.