Split History to EventReader/Writer as applicable
Namely, components that write to history (automator, release) take an
EventWriter, and components that read from history (server) takes an
EventReader.

This allows us to Tee the written history events to Slack, which will
pick up those events matching the regexp `Release:` and pipe them to the
configured Slack channel.