Use a database table to store automation state
This makes the simplest change necessary to persist the automation
flag for services. It simply writes the config to a database table
when it changes, and restores it when the automator starts.

NB that although the config database is created separately, it's
assumed (by the automator) that it will be used only by the
automator. Further design will be needed to make the automator able to
share the configuration database.