In-mem job store
- Create JobReadWriter (service) and JobPopper (worker) interfaces
- Create concept of JobSpec: Service/ImageSpec + Kind
- Create concept of Job: JobSpec + ID + status details
- Create in-memory implementation of JobStore
- Create release Worker, popping and executing jobs
- Unexport release.Release, as it becomes "owned" by the Worker

For now the execution is very coarse-grained. The worker claims the job,
invokes Release exactly as before, and updates the status/log with the
results only once it is completely finished.

In the next commit I will update fluxd to stop using Release directly,
and instead submit release jobs to the JobStore, as well as launching a
single release worker.

In a subsequent commit I will create a new GET /release endpoint,
connected to the JobReader side of the job store. I'll also update
fluxctl to block and poll the GET side until things are finished.

In a final commit, I will update the releaser to take a job object
instead of the spec, and continuously update it as it does its work.