Convert sync Release to async PutJob/GetJob
- Delete flux.Releaser
- Everything that used it now uses flux.ReleaseJobReadWriter
- Rather than calling Release, they call PutJob then GetJob in a loop
- This includes the Server and the Automator

Effectively code that looked like

  res, err := releaser.Release(service, image, kind)
  return res, err

Now has to look like

  id, err := releaser.PutJob(flux.JobSpec{service, image, kind})
  if err != nil {
      return nil, err
  }
  var res []flux.ReleaseAction
  for range time.Tick(time.Second) {
      j, err := releaser.GetJob(id)
      if err != nil {
          return nil, err
      }
      if j.IsFinished() {
          res = j.TemporaryReleaseActions
          break
      }
  }
  return res, nil

This is a very ugly commit because I had to avoid an import cycle
between package flux and package release. Namely, package flux contained
Server, which now took a release.JobReadWriter; but package release
defined JobSpec which contained flux.{Service,Image}Spec and flux.Kind.
The most expedient thing was to move the release types to package flux,
but I think this reveals a problem in our type layout: obviously types
related to the release stuff should live in package release. I'll try to
address this in a separate PR later.

Also, we bump the Go version used by Circle from the default (1.5) to
the new hotness (1.7.1) in order to get access to the math.Read function.