Releaser takes a Job object and updates it
- Keep utilizing ReleaseAction type
- Update job as part of ReleaseAction invocation
- Therefore always invoke RAs, but only call Do if Kind == Exec
- Update ReleaseAction results to be nicely human-readable.
- Remove ReleaseJob.TemporaryReleaseActions as they are no longer used

Since we no longer return a bunch of actions from the Release method,
instead actually going through and doing work and updating the job
status as we go, using ReleaseAction makes almost no sense at all. We
should *definitely* remove this type and the co-occurring logic, and
prefer to just do the relevant work (and update the job) inline. Modulo
ReleaseKindDryRun, of course.