platform/kubernetes.Cluster serializes operations
In the patented Bryan Boreham actor style, we launch a goroutine with
NewCluster which consumes closures from the actionc. Each exported
method is modified to send its work as a closure along the actionc. We
also add a stop method which closes the actionc; stopped Clusters cannot
be restarted. This design was chosen to minimize changes to the public
API and to clients.

In a subsequent commit I will change Regrade to operate on batches. And
in a final commit I will audit the release workflows to send regrade
batches to the platform as appropriate.