Introduce DB migrations and use for history DB
So far we have been able to simply create new tables. Now we need to
modify tables. This commit adds a `db` package, used from
fluxd/main.go, which makes sure the database is up-to-date.

The implicit requirement is that there are scripts in
db/migrations/{driver} for each update, for each supported driver.

This commit moves the creation of the history database into a
migration. Others can follow.

The `--database-*` arguments are collapsed into just the source, with
the driver being taken from that. `--database-driver` is deprecated --
removing it would mean that upgrades will not go smoothly, since the
new executable will be supplied an argument it's not expecting.