Return an error when an image can't be fetched
Previously, we wanted to tolerate the occasional failure to fetch
image metadata, since it was often merely informational.

However, the automater, and releases in general, rely on having image
times. Otherwise, they can flap (if the metadata for the latest image
happens to fail, the second latest gets promoted).

To avoid this, return early if there is any error fetching metadata,
and return the error. That way, the caller can decide whether it can
use the information or not.