Use "file:" and "memory:" to denote cznic/ql DBs
Most database/sql drivers register themselves such that the scheme of
the source URL is the name of the driver. cznic/ql, on the other hand,
expects either "file:" or "memory:" or unadorned paths as the source,
while the drivers are named "ql" and "ql-mem".

To have a uniform treatment using just a source URL, we need to either
translate the source URL if it's for a ql[-mem] driver, or look up the
driver name given the source URL. Since `file://...` is more familiar,
I've done the latter.

The `mattes/migrate` drivers work the other way around -- they are
looked up by scheme. Therefore the migration drivers for ql, included
here, are registered against "file:" and "memory:", but use the
drivers named "ql" and "ql-mem". Ooof!