Basic config command
This lets you upload a YAML file as your instance config. The entire
value is overwritten, so if you don't supply all the values, they may
be blanked.

Better semantics would be to only set the config items that are
provided; this would also open the way for setting individual bits of
config.

For getting the config, in the API there is an unimplemented "secrets"
option that should omit secrets like the git private key.

For setting the config, there is an unimplemented "clear" option,
which provides for the unsetting of configuration items (you provide a
config object with values in the slots you want to unset).