Split platform and fluxd from fluxsvc
We want to support daemons in many clusters with a single
service. This means splitting the components at `platform`; most
things are done in the service (fluxsvc), but querying and controlling
the platform is done from the daemon (fluxd).

This commit creates a new executable `fluxsvc`, and moves the
initialisation of components into the appropriate places.

It also introduces some infrastructure for communicating between the
daemons and the services, which will be improved later.

This involves some changes to the imageb builds, notably:

 * Build each image in its own directory
 * the daemon doesn't need git or python

Lastly, push both images to quay.io in the CircleCI script.