Add websocket-based RPC between fluxsvc and fluxd
This implements the platform interface via an RPC mechanism.

The flux daemon running in the instance cluster connects using a
websocket to the service; it then accepts requests, encoded as
JSON-RPC.

The flux service accepts incoming websocket requests. It treats each
websocket as the client side of an RPC connection. A `MessageBus`
provides the rendezvous between the server implementation and a
specific instance's connection.