Add Connecter and MultitenantInstancer
MultitenantInstancer is an implementation of Instancer that produces
instances by consulting the config database. It replaces
StandaloneInstancer. However, we cannot switch directly over for two
reasons:

 1. Systems using StandaloneInstancer will not have config record(s)
 containing the required settings, like git repo URL, so will break as
 soon as you try using them; and,
 2. We don't yet have a means of connecting to more than one platform.

The workaround for the first will be to deprecate the arguments for
StandaloneInstancer (`--repo-url` and so on), but not use
MultitenantInstancer until settings are in place.

The workaround for the second is to introduce an intermediate
interface Connecter, for establishing a connection to a platform
(agent). In standalone mode, an implementation can use a client to the
local Kubernetes cluster. In the future, the implementation can use a
rendezvous based on the instance ID or other configuration.