Rewrite to bounce on error, and use 1 subscription
The expected behaviour of Subscribe is that any errors will cause it
to exit (as well as being passed back to the caller), so do that.

There's no need to use a subscription for each method, when we have
wildcards. We have to switch on the method and do our own decoding,
but we had to do a (type) switch anyway. Doing it this way makes
tidying up easier.