Introduce IsDaemonConnected API endpoint
This adds a method `IsDaemonConnected` to the `DaemonAPI` interface,
for checking whether a particular daemon is known to be connected, and
exposes it in the HTTP transport.

It's not in the ClientAPI interface because I don't expect it to be
called from the command-line client (although that's somewhat arbitrary
at this point, anyway it can be moved later).

`IsDaemonConnected` asks the message bus, which checks for the
presence of a platform subscribed as the instance given. "Presence"
means that we can ping it via an RPC call -- necessitating a type
representing a platform that can be pinged, `RemotePlatform`.

Why ping? In some ways, just answering "yes" if there's a connection
registered would be neater and faster. However, we don't in general
know if connections are live, since they may not shut down cleanly --
and we only find that out after attempting to write to (and read from)
the socket. A secondary benefit of using a ping is that it
demonstrates that the process on the other side is still minimally
active.

(This also fixes a bug in which the test for cleaning up a connection
compared two interface pointers that would never be equal.)