Remove --secrets arg and never report secrets
This breaks the symmetry between SetConfig and GetConfig by using
distinct types to represent the instance config.

In the case of SetConfig, it accepts an `UnsafeInstanceConfig`, that
is one with secrets exposed (we need to do this because otherwise you
can't supply, e.g., registry credentials). In the case of GetConfig,
it returns a regular InstanceConfig. The difference between the types
is that UnsafeInstanceConfig obscures secrets before serialising them
to JSON, for transmission.

We have to be a little careful, since we *also* serialise when
outputting to the console, and when storing in a DB -- these
operations must use UnsafeInstanceConfig so that they serialise
faithfully.