Correct lookup of "library" images
Some image names do not include an org; e.g., "nats:0.8.0". These are,
by convention, kept on docker hub in the org "library". However, in
general we want to deal with image names as they are supplied --
`nats:0.8.0` rather than `library/nats:0.8.0`.

The convention only really matters when looking up images in the
registry; so, this commit simply changes the code there to distinguish
between the name used to look up the image and the name to be used in
the result.