Make fluxsvc regrade async, gracefully stop
When upgrading fluxsvc itself, we effectively ask Kubernetes to
restart the pod(s) that's watching the regrade process, and thereby
lose the result.

This commit makes `fluxsvc` a special case which is run
asynchronously. That is, the regrade is started but the service does
not wait for it to complete before reporting back to the user.

There is still a race between reporting back to the user, and the pod
being restarted. For that reason, the service attempts to shut down
gracefully, waiting until all release processes are completed, with a
generous timeout. This means the release procedure has a chance to
record what happened.

This commit also changes the worker loop so that it polls for the next
job with a backoff, rather than polling on a tick.