Remove semaphore from releaser
The releaser used a channel as a mutex, allowing only one release to
happen at a time. When there was one fluxsvc and one platform in
operation, this made some sense.

However, we now have multiple platforms (one per instance), and want
to run more than one fluxsvc. Having a mutex guarding multiple
unrelated platforms is counter-productive; and anyway irrelevant if
there are multiple running services.

Regrades are still serialised in the platform, so there is still a
guard against running interacting regrades.