Lengthen timeout for Regrade
The Regrade method can take an arbitrary amount of time, since it
depends on k8s completing a rolling upgrade. While we stick to the
current daemon-service protocol, this means we can't fail a regrade
after 5 seconds -- most of them will take longer than that.

So, set it to twenty minutes, because why not. The risk that comes
with this is that a particular call *should* time out after five
seconds, because the platform has actually gone away and will never
answer. However, since Regrade comes after other calls (on short
timeouts) to the platform, there's a fairly small window for this to
happen.