Stop outputting kubeservice traces
There's seldom need to look at the output, which ultimately tells you
very little except that a lot of files weren't the file you were
looking for.

On the other hand, it is a large proportion of the logs, and is
frequently mistaken for indicating a bug (where the apparent failures
are just because of the quirky way flux uses kubeservice).