Change slack hook regexp
The aim is to catch the concluding message. Using `"Release"` will
catch the `"Starting ..."` messages, since they now include that word
instead of `"Regrade"`, while missing the `"Apply ..."` messages.

Using `"done|failed"` is still fragile, but does the job. (NB it
required changing the ordering of the failure message, so that
"failed" is always last.)