Create a binary to query a SQL db and expose for prometheus.
This adds a simple binary that connects to an Postgres-like DB (e.g. RDS), performs a set of user-specified queries and exposes them on the `/metrics` endpoint.

The binary is added to the fluxsvc dockerfile. To use, alter the entrypoint to call the `fluxmon` application.

Queries are specified in a yaml file.

See `--help` for all options.