Rewrite releaser code
The releaser code works by accumulating a list of "release action"
structures, then running through them.

But the actions mostly do not correspond with what we want to tell the
user about -- most of them are just preparation (e.g., "clone the
repo"), which does not have a place in the plan we report to the
user. This means the useful information is interleaved with irrelevant
details, and decisions made during the process (e.g., about which
services to ignore) are obscured.

A better structure is to simply do all the preparation, which is for
the most part the same each time, in one place, while keeping an
explicit record of the decisions about what to do in the execution
phase.

This commit also changes the planning code so that git and the
instance config are examined before the running system. This means we
don't waste time (and the user's attention) on services that can never
be affected.

Now that the release is more straight-forward, the automator can be
rewritten in the same terms.

NB this commit involves a change to the platform (and therefore
fluxd):

Previously a release with `--all` would get all the services running
in the platform, then see which of them is defined in the files. Now,
we get all the services defined in the files, and ask the platform
about them specifically.

This exposes a problem with the platform implementation: instead of
skipping services that are requested but not present, it returned an
error. Before, this would only occur if the user specifically
mentioned the (missing) service, and they would get back the error
saying it's missing, which all seemed reasonable. Now, a user will get
the error without having done anything "wrong".

The correct (and documented :/) behaviour is to simply miss out the
missing service, and let the service figure out that it was defined,
but not running.