Treat missing services as skipped if requested
If a service is defined in the repo but not running in the platform,
we mark it as `ignored`. However, if the service was specifcially
requested (e.g., with `--service=...`), this may leave the output
blank, since ignored services aren't printed by default.

So, if a service was explicitly requested but is missing, mark it as
`skipped`, so it always shows up in output.