Avoid a loss of fatal errors in subscriber
In the NATS bus subscriber, all errors resulting from platform
operations are forwarded to an error channel, which is checked by a
loop. The receiving loop checks each error from the channel to see if
it's a FatalError, which must shut the subscription down.

If the channel send (by the individual operation is blocked, the error
is discarded, on the assumption that there's some other error being
processed. But, because *any* error can be blocking the channel, that
means a FatalError can be discarded in favour of another non-fatal
error.

The solution is to only send FatalErrors to the channel.