Use git output in errors
At the minute, if the git commands (in particular, clone and push)
fail, the user is presented with the error from os/exec.Run() which is
more or less an opaque exit code.

To be a bit more informative, we can instead take the output to stderr
and look for a "fatal:" line, which generally gives a compact
description of what actually went wrong.