Erect scaffold for returning humane errors
We would like to take an incremental approach to improving the errors
returned via the API. For that reason, we have to deal with any old
error coming through, while doing better where we can.

The approach taken here is to test for some variety of flux.BaseError,
and encode into the HTTP response accordingly. This will mostly
require a new version for each route, since the contract at present is
to use HTTP status 500 for almost any error, and to simply write the
error text into the response.

The humane errors come with help text which can be printed out
verbatim (most use for the CLI), as well as a symbolic name and
particulars that can be used to generate markup (most use for web
UI). Lastly, there's the original error as a string, in case this
needs to be logged.

This commit also gives an example of returning a humane error from an
API call (for a missing release job).