Adapt API version test to use HTTP status code 410
HTTP 404 Not Found was previously assumed to mean that the client had
called a route that is no longer (or not yet) supported by the
service. However it's now used to indicate that a particular resource
is missing, e.g., a job.

The equivalent for the purpose of the test is 410 Gone, meaning an API
has been deprecated.

This adds routes to the API corresponding to deprecated API versions
(v1 and v2) represented by 410 Gone, and a last resort 404 Not Found
for anything that doesn't match.

This means no special handling in the client, since it will by default
receive the error and print it out. Old clients may see a 404 in
places, but they are wired to treat that as a version mismatch anyway.

The test for version mismatch is no longer needed, since there's no
special treatment of those results.