Store job errors in the job table
To be able to output a helpful error message to the user when a job
(e.g., a release) fails, we need to record the error. To this end,
this commit adds an error field to each job, and a corresponding
column to the database table.

To avoid difficulty in encoding and decoding generic errors, it's
assumed that the error is a `flux.BaseError`; i.e., an error with help
text.

This commit also condenses some of the job encoding and db code.