Periodically boot web socket connections
For usage stats, we report on HTTP transactions; and websocket
connections are effectively long-running HTTP transactions. Naively,
if a websocket is open over several days, it will get counted as a
"use" only at the end, so only once.

To get more accurate stats, we can limit the age of websocket
connections, so connections are reported more often. The limit will of
course affect how many transactions are reported; however, the thing
we generally care about is _distinct_ connections, i.e., the number of
daemons connected. A limit of two hours is somewhat arbitrary --
anything (much) less than a day would do.

The disconnection is done on the application layer (in bus.go), so
that it doesn't happen in the middle of an operation. Exiting there
prompts the server to close the RPC client, which itself closes the
underlying websocket, to which the daemon responds by reconnecting
immediately.