Remove two unnecessary close(chan) calls
The channels in question will be garbage collected, and do not have
loops ranging over them.

In some situations, it is dangerous to do this, since a goroutine may
be waiting to send on the channel and leak; closing it would at least
expose that problem. In this case, the sends are in a select with a
default branch, so will not block.