Deprecate updating ReplicationControllers
There are a few places where we can detect replication controllers and
return an error:

 * when listing the services or images
 * when updating a manifest
 * when applying an updated manifest

In the first case, I thought it would be strange to remove information
from the UI by supplying an error instead of container information. So
that stays as it is.

The second case is the best point, since it occurs just before a
lasting change (git commit) is made. The third case, failing to apply
the change, is really a backstop -- a change has been committed by
this point, which shouldn't have happened -- but it means we can
remove a bunch of code dealing with `kubectl rolling-upgrade`.

There is a choice between failing to update _just this service_, and
failing a release altogether. I have opted for the latter, because it
means we can put a big help message in front of the user.

(A more subtle (but maybe confusing?) approach is to fail just the
service, if it's one of many, or fail the whole thing if the release
is specifically for this service.)