Switch to {Apply,Delete} for sync actions
These are declarative ("make it like this") rather than operational
("do this"), which means more latitude for the implementation to do
useful things.

Since the purpose of Sync is to make the resources in a cluster match
those in a repo, it makes more sense to say what the desired state is,
rather than how to get there. For example, Apply can act like
"upsert", creating a resource if it doesn't exist.