Added HTTP API to return the value of a single configuration key
The new API is called GetSingleConfig can be found at GET /v5/config?key={key}&syntax=[yaml|json].
Much of the new logic is located in the config.go file. The goal is to reflectively iterate and compare every configuration key to that of a requested configuration path format (e.g. git.key). The key can follow the yaml keys or the json keys using the syntax parameter. The default is yaml. (Currently both syntaxes are the same).
The options that have periods must be quoted in single quotes.
The options use the yaml tag in the configuration struct to match the get-config names.