Implement sync
This requires a bit of scaffolding to be able to compare the running
environment (which is obtained by cluster.Export) with the configured
environment (loaded from the files).

I've deliberately put a bit more structure than strictly necessary, in
anticipation of using this for other purposes, e.g., replacing
`kubeservice`.