Encode test invocation into Makefile
This moves the invocation of `go test` into the Makefile, for
convenience of users and the CircleCI script; and removes the need for
`glide` to be installed.

NB drops the `go build ...` step from the CircleCI test stage -- it's
not doing anything that `go test` and `make all` won't do, and it adds
significantly to the build time.

Also drops the invocation of `go test` without the tag `integration`,
since those get run when the tag's present too.