If `fluxctl --all --update-image=xxx` only show services that pertain to that image
Most of the code are tests to verify the following requirement:
```
iff (-s=service-name || (--all && --update-image=xxx)) ;
   only services that pertain to an image should be shown in the output
```
Added logic to satisfy the part (--all && --update-image=xxx).
If this is true, then it manually deletes all keys in the result that do not pertain to the image.
This removes the services that do not pertain to the current image and fulfils the requirement.
Note that the intial requirement is already satisfied by the image search. But a test was added to explicitly test the requirement.

Fixes #492