Rationalise included/excluded/locked status.
Implemented `ServiceFilter` to filter defined or running services.
The service is ignored if: excluded OR not included OR not correct image.
The service is skipped if: not ignored AND (locked or not found in cluster)
Else the service is pending.
The filters must be run across both the services to update and the services that are not running in the cluster. This is because the services not running in the cluster might have been ignored anyway.

Fixes #492