Separate the cluster and platform packages
It used to be the case that platform.Platform was both the interface
to the running cluster _and_ the interface exposed by the daemon to
the service over RPC.

Now that most operations are done in the daemon, these are quite
distinct, and keeping them together makes it difficult to define new
packages that depend on the `Cluster` interface without dragging in
platform (and introducing import cycles that then need to be worked
around).

This commit extracts the Cluster interface, and its implementation for
kubernetes, into new packages, and renames `platform` to `remote`, to
describe its purpose better.