Keep a tag pointing to the sync'd revision
Presently this points at the last revision that was
synchronised. There are some problems with this design:

 * If it's never been synchronised, the tag will not exist (and
   sync-status will fail)

 * If the daemon has just started up, it will have no idea how out of
   sync the cluster is, limiting the usefulness of the tag

Were sync to be running continually, rather than triggered, these
become less of an issue, since the tag will generally be present and
close to recent while the daemon is there to answer.

(In fact, doing a sync on starting up would also mitigate these
problems)