Bring the bus implementation of Ping up-to-date
The various platform.Platform implementations of Ping return a
platform.UnavailableError if the call fails to return (including if
the NATS request times out). However, the method actually used by the
server is MessageBus.Ping -- the platform.Platform implementation is
side-stepped.

This commit makes the error handling for NATS.Ping behave the same as
for natsPlatform.Ping.