Keep notes, use a working clone for mutating files
This introduces an API for dealing with `Checkout`s, which represent a
mutable local clone of the (config) repo. This is to tie operations to
a particular working directory, and to make it easier to create
short-lived clones when forming new commits.

The CommitAndPush method now includes a possibility of writing a
note (per `git notes add`) with the semantic result of an operation,
e.g., a release. The notes are pushed, and pulled, along with the
commits themselves. The idea is to make the results part of the
"source of truth", so they can be consulted when sync'ing later.