Account for failure to dial TCP too
When trying to establish the websocket connection it's possible for
the TCP connection to fail, e.g., if the URL has hostname that can't
be resolved. In this case, just pass the error on rather than
constructing a `DialErr` (which presumes it got an HTTP response, and
will NPE if it didn't).