Change sync API to Notify/Status
This changes the `SyncCluster` method to `SyncNotify`, adapting its
meaning to be "consider syncing now".

Other things are also simplified: sync/ now just syncs a cluster to a
directory of files.

Lastly, the sync/automation Loop is now run in main.go, so the cluster
will *actually* synchronise with the repo; I have hedged slightly by
adding a `delete` argument (delete things if not present) and fixing
it to false for now.