Run releases and policy updates through job queue
To serialise operations on the repository, things that mutate files
are run through a jobs queue. This involves changing the signature in
remote.Platform methods to return a job ID, so they can return more or
less straight away (which is the ideal for these RPCs).

In the daemon's loop, each job is run in turn. When successful, a job
is assumed to have psuhed commits, so a sync is triggered.

For now I have changed the

    fluxctl release|lock|unlock|automate|deautomate

commands to simply return the job ID. The next step is to get them to
poll for the state of the job (and to keep recent failures around, so
they can be reported on).