Conflate Image and ImageDescription
These two types have the same fields and are put to the same use, so
just use one of them everywhere.

Also, use a time.Time value rather than a pointer, and fall back to
time.IsZero() to detect missing/uninitialised values. Although it's
nice to have a nullable value and rely on the JSON encoder to miss it
out or serialise it to `null`, it is also a repeating source of
confusion, since it's unusual to use a pointer to a value type.

The custom JSON marshalling is not strictly necessary (a zero time
value will roundtrip), it's tidier and matches the expectation from
clients of the API.

The latter is also the reason for naming the ID field: in the API it
is serialised as `ID`, so it is simpler to just call it that
everywhere.