Use port 3030 for the daemon API
You used to run the daemon and service in the same pod, meaning they
had to listen on different ports. But that's no longer necessary, in
fact it's no longer necessary to run the service at all. And it's less
confusing if the API is consistently at the same port.