Remove obsolete fluxctl commands
This removes the commands that do not correspond to the daemon
API. They are

 * history (now supported just by the service)
 * get- and set-config (config relevant to the daemon is supplied as
   args)
 * status (useful for the service, less so the daemon)

The last one is a borderline case; we might usefully make a similar
status response for the daemon. But I'm removing it until we decide
what the requirement for that is, if any.

This also adds Export to the daemon API, so `fluxctl save` will work
when targetting the daemon directly.