Change the job queue contract
Formerly it was possible to get an arbitrary number of `nil`s from the
job queue once you shut it down, since it closed the ready channel
when stopping.

Closing the ready channel was intended to let any goroutine that was
blocked on receiving from it proceed. Unfortunately, when used in a
select loop, it means that case will always succeed; and to avoid
problems, consumers of the queue had to be shut down synchronously
before it was shut down.

This commit makes the ready channel always block, on the understanding
that it will be used in a select block with at least one other case
that will eventually succeed. This means the queue and its consumers
have no required shutdown order.

It also removes a similar attempt to unblock goroutines trying to
enqueue items -- this would never work in general, since it only
received one value!

For the purpose of testing, I added a `q.Sync()`, with heavy
caveating.