Add daemon tests
Coverage is at about 80%. The parts that are not covered are error cases
(e.g. job failed, etc.).

The git repo being used during the test is real. Hence, any tests that
alter manifests really will alter files.

The k8s cluster used in the daemon is a mock. Hence, this doesn't
actually test whether a new update is applied to the cluster.

The daemon is quite complex to set up. The key thing is stopping the
jobs queue loop and the daemon loop in a specific order. If the queue
stops first, the daemon will receive an invalid nil job.