Explicitly model "not ready" daemon state
To facilitate configuration of a generated keypair as a deploy key via
the UI, it is necessary to move the initial git clone operation after
the upstream connection, and to retry it periodically until it is
successful. This commit introduces a NotReadyDaemon which is capable of
supporting specific operations critical to the task (e.g. getting the
public key) but not others which depend on the cloned config repository;
once the full daemon is available, it is substituted in its place.