Simplify release event structure and include cause
Previously an update.Release{} was constructed and populated to send
as part of a release event, and that struct was used only for that
purpose. But there's no reason to nest another struct, since there's
already the single-purpose `history.ReleaseEventMetadata`, so this
commit flattens that out.

It also makes sure that the cause is included with release events, so
that history and notifications are accurate.

I have taken no care to make this backwards compatible with previous
git notes or history entries; but since what is shown in the UI is
just the summary string (which is computed before storing), I don't
think this will matter much if at all.