Make loop request a sync if pushing fails
If there's commits in the upstream repo that have not made their way
to the local system, pushing will fail. In this case, it's a
reasonable tactic to return an error while also requesting a sync,
which will fetch those commits. If the operation is attempted again,
it has more chance of succeeding with the newly-pulled repo.

There are other reasons that the commit-and-push can fail; in those
cases, requesting a sync is harmless (it will just sync sonner than it
would have).