Rewrite sync loop to aid short-cuts
There's two short-cuts we may wish the loop to take.

 1. If pushing commits fails, it is likely to be because it was not
 possible to fast-forward. In this case we'd like to pull straight
 away so that any subsequent attempt at the job is more likely to
 succeed.

 2. If a job switches automation on for something, it would be good to
 run the image poll straight away to appear responsive to the change.

In the latter case, it is easy enough to set up a "poll images soon"
mechanism similar to `d.askForSync()`.

A subtlety is that there is then a race between the commit just made
(automating something) and automating running -- if the latter wins,
it will ignore the newly automated thing because it won't see the
commit that makes it so. To avoid that, the control flow is altered so
that when a job is successful, the repo is pulled from upstream before
returning to the loop.