Allow for possibility of no note on a commit
`git note show` will exit with 1 if there is no note for a given
commit (including the case of the notes ref being unknown). We want to
distinguish between this and there being a genuine error reading a
note: so, return `(nil, error)` if there was a problem, and `(nil,
nil)` if there was simply no note.

This needed better git error message parsing, so I cherry-picked a
commit doing just that from `resync-on-push-fail`.