Propagate helpful errors through RPC
Sending errors over RPC is delicate. The scheme in `net/rpc` is that
any error returned by the server is stringified, and made into a new
error at the client end.

Since we want to preserve the structure of _application_ errors
returned by the daemon, we must embed them in the response field. This
is a lot easier to do if there's a single type -- so one big change in
this commit is that different kinds of application error are
discriminated using a field of a single Error type, rather than being
a type each.

They also need transmission over the bus, so there's a compensating
change to the bus code.

Because this fundamentally changes the data sent over RPC, it needs a
new protocol version. (These still share version numbers with the API,
since the connection endpoint -- RegisterDaemon -- is part of that
API).